+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fix sitemaps namespace
 - Update livereload
 - Add `hard_link_static` config option to hard link things in the static directory instead of copying
+- Draft pages are not longer rendered
 
 ## 0.8.0 (2019-06-22)